+++ client/src/scripts/types/InputField.ts
@@ -13,6 +13,7 @@ interface InputField extends GarnishComponent {
  */
 interface NeoInputField extends InputField {
   getBlocks: () => NeoInputBlock[]
+  getName: () => string
   removeBlock: (block: NeoInputBlock) => void
 }