+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 ### Added
+- Added the `barsDisallowedFields` plugin setting, for setting Matrix or Neo field handles for which the batch actions bar won't be initialised
 - Added `spicyweb\batchactions\assets\bars\BarsAsset`
+- Added `spicyweb\batchactions\models\Settings`
 
 ### Deprecated
 - Deprecated `spicyweb\batchactions\assets\BatchActionsAsset`; replaced with `spicyweb\batchactions\assets\bars\BarsAsset`